[BPK-4276] Split up A-web component pages
